version 1.0

+++ src/sortByRatio.java
@@ -2,7 +2,7 @@ import java.util.Comparator;
 
 public class sortByRatio implements Comparator<Item> {
     public int compare(Item a, Item b) {
-        boolean temp = (float) a.getUtility() / a.getWeight() > (float) b.getUtility() / b.getWeight();
+        boolean temp = a.getUtility() / a.getWeight() > b.getUtility() / b.getWeight();
         return temp ? -1 : 1; // descending order
     }
 }
